Fix broken path to textures and soundfonts
I never noticed it before because I pointed the web server root to the repository root, so ../../ had the same result as ../ from within the src directory. However, it makes a difference when the web server root is at a higher level and trying to retriece the textures and soundfonts would lead to an NOT_FOUND HTTP error. The proper path now uses ../ instead of ../../ .