Work around connections problems of leap.js to leapd
For some reason, leap.js sometimes can not properly (re-)connect to the leapd background service and this causes the app to freeze. The problem seems to be related to leap.js internal handling of automatic re-connects. Therefore, we disable automatic reconnects via leap.js and move the high-level connections handling out of the library into the app code.